requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
httpx
redis
aioredis
bson
pydantic_settings